+++ android/TestJunit5/src/test/resources/conf/parallel.conf.json
@@ -1,0 +1,21 @@
+{
+  "server": "hub-cloud.browserstack.com",
+  "user": "darshilgada1",
+  "key": "fP6DdHuqEjCUEmnvJmz7",
+
+  "capabilities": {
+    "project": "Test Junit5",
+    "build": "Test junit-browserstack",
+    "name": "parallel_appium_test",
+    "browserstack.debug": true,
+    "app": "bs://2e6540b05c17bb948bd72baf3f4f5e33948ec415"
+  },
+
+  "environments": [{
+    "device": "Google Pixel"
+  }, {
+    "device": "Samsung Galaxy S7"
+  },{
+    "device": "Samsung Galaxy S20"
+  }]
+}
