+++ android/junit4/src/test/resources/com/browserstack/run_single_test/single.conf.json
@@ -1,0 +1,18 @@
+{
+    "server": "hub-cloud.browserstack.com",
+    "user": "darshilgada1",
+    "key": "fP6DdHuqEjCUEmnvJmz7",
+  
+    "capabilities": {
+      "project": "Test Junit4",
+      "build": "Test junit-browserstack",
+      "name": "single_appium_test",
+      "browserstack.debug": true,
+      "app": "bs://2e6540b05c17bb948bd72baf3f4f5e33948ec415"
+    },
+  
+    "environments": [{
+      "device": "Samsung Galaxy S7"
+    }]
+  }
+
