+++ android/junit5/src/test/java/com/browserstack/run_parallel_test/BrowserStackJUnitTest.java
@@ -1,0 +1,95 @@
+package com.browserstack.run_parallel_test;
+
+import com.browserstack.local.Local;
+import io.appium.java_client.android.AndroidDriver;
+import io.appium.java_client.android.AndroidElement;
+import org.json.simple.JSONArray;
+import org.json.simple.JSONObject;
+import org.json.simple.parser.JSONParser;
+import org.json.simple.parser.ParseException;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+
+import org.junit.jupiter.api.parallel.Execution;
+import org.junit.jupiter.api.parallel.ExecutionMode;
+import org.openqa.selenium.remote.DesiredCapabilities;
+
+import java.io.FileReader;
+import java.io.IOException;
+import java.util.*;
+import java.util.stream.Stream;
+
+@Execution(ExecutionMode.CONCURRENT)
+public class BrowserStackJUnitTest {
+
+    public AndroidDriver<AndroidElement> driver;
+    private Local l;
+    public String username;
+    public String accessKey;
+    public DesiredCapabilities capabilities;
+
+    public static JSONObject config;
+
+    private static Stream<Integer> devices() throws IOException, ParseException {
+
+        List<Integer> taskIDs = new ArrayList<Integer>();
+
+        if(System.getProperty("config") != null) {
+            JSONParser parser = new JSONParser();
+            config = (JSONObject) parser.parse(new FileReader("src/test/resources/com/browserstack/run_parallel_test/" + System.getProperty("config")));
+            int envs = ((JSONArray)config.get("environments")).size();
+
+            for(int i=0; i<envs; i++) {
+                taskIDs.add(i);
+            }
+        }
+
+        return taskIDs.stream();
+    }
+
+    @BeforeEach
+    public void setup() throws Exception {
+        System.out.println("@BeforeEach executed");
+
+        capabilities = new DesiredCapabilities();
+
+        Map<String, String> commonCapabilities = (Map<String, String>) config.get("capabilities");
+        Iterator it = commonCapabilities.entrySet().iterator();
+        while (it.hasNext()) {
+            Map.Entry pair = (Map.Entry)it.next();
+            if(capabilities.getCapability(pair.getKey().toString()) == null){
+                capabilities.setCapability(pair.getKey().toString(), pair.getValue().toString());
+            }
+        }
+
+        username = System.getenv("BROWSERSTACK_USERNAME");
+        if(username == null) {
+            username = (String) config.get("user");
+        }
+
+        accessKey = System.getenv("BROWSERSTACK_ACCESS_KEY");
+        if(accessKey == null) {
+            accessKey = (String) config.get("key");
+        }
+
+        String app = System.getenv("BROWSERSTACK_APP_ID");
+        if(app != null && !app.isEmpty()) {
+            capabilities.setCapability("app", app);
+        }
+
+        if(capabilities.getCapability("browserstack.local") != null && capabilities.getCapability("browserstack.local") == "true"){
+            l = new Local();
+            Map<String, String> options = new HashMap<String, String>();
+            options.put("key", accessKey);
+            l.start(options);
+        }
+    }
+
+    @AfterEach
+    public void tearDown() throws Exception {
+        System.out.println("@AfterEach executed");
+        driver.quit();
+
+        if(l != null) l.stop();
+    }
+}
