+++ android/junit5/src/test/java/com/browserstack/run_parallel_test/ParallelTest.java
@@ -1,0 +1,49 @@
+package com.browserstack.run_parallel_test;
+
+import io.appium.java_client.MobileBy;
+import io.appium.java_client.android.AndroidDriver;
+import io.appium.java_client.android.AndroidElement;
+import org.json.simple.JSONArray;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.MethodSource;
+import org.openqa.selenium.support.ui.ExpectedConditions;
+import org.openqa.selenium.support.ui.WebDriverWait;
+
+import java.io.IOException;
+import java.net.URL;
+import java.util.Iterator;
+import java.util.List;
+import java.util.Map;
+
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+public class ParallelTest extends BrowserStackJUnitTest {
+
+  @ParameterizedTest
+  @MethodSource("devices")
+  void testCalcOne(int taskId) throws IOException, InterruptedException {
+
+    JSONArray envs = (JSONArray) config.get("environments");
+
+    Map<String, String> envCapabilities = (Map<String, String>) envs.get(taskId);
+    Iterator it = envCapabilities.entrySet().iterator();
+    while (it.hasNext()) {
+      Map.Entry pair = (Map.Entry)it.next();
+      capabilities.setCapability(pair.getKey().toString(), pair.getValue().toString());
+    }
+
+    driver = new AndroidDriver(new URL("http://"+username+":"+accessKey+"@"+config.get("server")+"/wd/hub"), capabilities);
+
+    AndroidElement searchElement = (AndroidElement) new WebDriverWait(driver, 30).until(
+            ExpectedConditions.elementToBeClickable(MobileBy.AccessibilityId("Search Wikipedia")));
+    searchElement.click();
+    AndroidElement insertTextElement = (AndroidElement) new WebDriverWait(driver, 30).until(
+            ExpectedConditions.elementToBeClickable(MobileBy.id("org.wikipedia.alpha:id/search_src_text")));
+    insertTextElement.sendKeys("BrowserStack");
+    Thread.sleep(5000);
+
+    List<AndroidElement> allProductsName = driver.findElementsByClassName("android.widget.TextView");
+    assertTrue(allProductsName.size() > 0);
+
+  }
+}
