--- android/src/test/resources/conf/local.conf.json
@@ -1,17 +1,0 @@
-{
-  "server": "hub.browserstack.com",
-  "user": "BROWSERSTACK_USERNAME",
-  "key": "BROWSERSTACK_ACCESS_KEY",
-
-  "capabilities": {
-    "build": "junit-browserstack",
-    "name": "local_appium_test",
-    "browserstack.debug": true,
-    "browserstack.local": true,
-    "app": "bs://<hashed app-id>"
-  },
-
-  "environments": [{
-    "device": "Samsung Galaxy S7"
-  }]
-}
