--- android/src/test/resources/conf/parallel.conf.json
@@ -1,20 +1,0 @@
-{
-  "server": "hub.browserstack.com",
-  "user": "BROWSERSTACK_USERNAME",
-  "key": "BROWSERSTACK_ACCESS_KEY",
-
-  "capabilities": {
-    "build": "junit-browserstack",
-    "name": "parallel_appium_test",
-    "browserstack.debug": true,
-    "app": "bs://<hashed app-id>"
-  },
-
-  "environments": [{
-    "device": "Google Pixel"
-  },{
-    "device": "Samsung Galaxy S7"
-  },{
-    "device": "Google Nexus 9"
-  }]
-}
