--- android/src/test/resources/conf/single.conf.json
@@ -1,16 +1,0 @@
-{
-  "server": "hub.browserstack.com",
-  "user": "BROWSERSTACK_USERNAME",
-  "key": "BROWSERSTACK_ACCESS_KEY",
-
-  "capabilities": {
-    "build": "junit-browserstack",
-    "name": "single_appium_test",
-    "browserstack.debug": true,
-    "app": "bs://<hashed app-id>"
-  },
-
-  "environments": [{
-    "device": "Samsung Galaxy S7"
-  }]
-}
